fix(api): todos os ajutes feitos para remoção da tabela moveset e adição da favorites
