+++ api/app/Http/Requests/know_area/StoreKnowAreaRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\know_area;
 
 use Illuminate\Foundation\Http\FormRequest;
