+++ api/app/Http/Requests/topic/StoreTopicRequest.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\topic;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateFavoriteRequest extends FormRequest
+class StoreTopicRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
      */
     public function authorize(): bool
     {
-        return false;
+        return true;
     }
 
     /**
@@ -22,7 +22,7 @@ class UpdateFavoriteRequest extends FormRequest
     public function rules(): array
     {
         return [
-            //
+            "name" => ["required", "string", "max:255", "unique:topics"],
         ];
     }
 }
