+++ api/app/Http/Requests/topic/UpdateTopicRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\topic;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreSpecialtyRequest extends FormRequest
+class UpdateTopicRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,7 +22,7 @@ class StoreSpecialtyRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'name' => ["required", "string", "max:255", "unique:know_areas"],
+            "name" => ["required", "string", "max:255", "unique:topics"],
         ];
     }
 }
