--- api/app/Models/Favorite.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class Favorite extends Model
-{
-    /** @use HasFactory<\Database\Factories\FavoriteFactory> */
-    use HasFactory;
-}
